Create README.md
This project was completed during Hackfest 2019, at IIT(ISM) Dhanbad. The project aims at developing a sound recognition system for the deaf people. A microphone recieves the input voice signal. Then this voice sequence is converted to sequences of 5 bit binary numbers. We use LSTM encoder decoder architecture based model for this sequence to sequence modelling. Each 5 bit block in the obtained translated sequence is used to vibrate 5 piezoelecric sensors attached to hand. Thus, a deaf person who is trained in the language of vibrations to understand the pattern recognises the message.
The advantage of using a DL model and not hard coding the words is that the user just has to learn only the vibratory language. And for any natural language the model can be trained to convert to that sequence. Thus same embedded device can be made more robust, as we care about the actual meaning inside the sound recieved, to obtain the vibrations.